batch/requirements.txt
asciitree==0.3.3
astmonkey==0.3.6
astroid==2.15.8
black==23.9.1
bytecode==0.14.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.2
docstring-parser==0.15
exceptiongroup==1.1.3
idna==2.8
iniconfig==2.0.0
jellyfish==1.0.1
Jinja2==3.1.2
lazy-object-proxy==1.9.0
libcst==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
MutPy-Pynguin==0.7.1
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
networkx==3.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.21.12
pydot==1.4.2
Pygments==2.16.1
pynguin==0.34.0
pyparsing==3.1.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
simple-parsing==0.1.4
termcolor==2.3.0
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.25.11
wrapt==1.15.0
